web/main/wsgi.py
"""
WSGI config for main project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/dev/howto/deployment/wsgi/
"""

import os
import dotenv

from django.core.wsgi import get_wsgi_application

ENV_PATH = os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env')

dotenv.load_dotenv(ENV_PATH)

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'main.settings')

application = get_wsgi_application()
