requirements.txt
absl-py==1.4.0
altair==4.1.0
appnope==0.1.3
argcomplete==1.12.0
argon2-cffi==21.1.0
arrow==1.0.3
asgiref==3.6.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
Beaker==1.10.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
bleach==4.1.0
blinker==1.4
blis==0.7.5
bs4==0.0.1
cached-property==1.5.1
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.10.8
cffi==1.14.1
cftime==1.5.1.1
chardet==3.0.4
charset-normalizer==2.0.8
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cookiecutter==2.1.1
cryptography==39.0.1
cycler==0.10.0
cymem==2.0.6
dasbus==1.4
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
deluge==2.0.3
distro==1.5.0
Django==4.1.5
django-admin-sortable==2.2.4
django-djconfig==0.10.0
django-haystack==3.2.1
django-infinite-scroll-pagination==1.1.0
django-leaflet==0.28.3
django-spirit @ git+https://github.com/GeorgeKandamkolathy/Spirit.git@d2abb44ee4ad03d1a3318e13b0c02edcd2b948a6
docx==0.2.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
flatbuffers==23.1.21
fonttools==4.38.0
fpdf2==2.6.1
ftfy==6.0.3
gast==0.4.0
# GDAL @ file:///C:/Users/tcf19/OneDrive/Documents/Orefox%20Employment/github/project_orefox_d4/web/GDAL-3.3.3-cp38-cp38-win_amd64.whl
gitdb==4.0.8
GitPython==3.1.24
google-api-core==1.34.0
google-api-python-client==2.79.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==2.16.1
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.48.2
gssapi==1.6.9
gunicorn==20.0.4
h11==0.14.0
h5py==3.8.0
httplib2==0.21.0
humanize==0.5.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.12.0
incremental==17.5.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.11.0
kiwisolver==1.3.2
koji==1.26.1
langcodes==3.3.0
langtable==0.0.56
leaflet==0.0.3
libclang==15.0.6.1
lxml==4.5.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.3.4
matplotlib-inline==0.1.3
matplotlip==0.2
Mesa==0.8.8.1
mistune==0.8.4
murmurhash==1.0.6
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netCDF4==1.5.8
networkx==2.5.1
nltk==3.6.7
notebook==6.4.5
numpy==1.23.4
oauthlib==3.2.2
olefile==0.46
opencv-python==4.5.4.58
openpyxl==3.1.1
opt-einsum==3.3.0
outcome==1.2.0
packaging==20.9
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
Paste==3.4.0
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==8.4.0
ply==3.11
porter2stemmer==1.0
poyo==0.5.0
preshed==3.0.6
productmd==1.33
progressbar2==3.51.4
prometheus-client==0.11.0
prompt-toolkit==3.0.21
proto-plus==1.18.1
protobuf==3.19.6
psycopg2==2.9.5
ptyprocess==0.6.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pycparser==2.20
pycurl==7.43.0.5
pydantic==1.8.2
pydeck==0.7.0
pyenchant==3.1.1
Pygments==2.10.0
PyHamcrest==1.9.0
pykickstart==3.29
Pympler==0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.8
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==0.21.0
python-slugify==4.0.1
python-utils==2.4.0
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
pyudev==0.22.0
pywin32==305
pywinpty==2.0.10
pyxdg==0.26
PyYAML==6.0
pyzmq==22.3.0
regex==2022.10.31
rencode==1.0.6
requests==2.28.2
requests-file==1.4.3
requests-ftp==0.3.1
requests-gssapi==1.2.2
requests-oauthlib==1.3.1
rsa==4.7.2
scikit-learn==1.0
scipy==1.10.1
selenium==4.8.2
Send2Trash==1.8.0
service-identity==18.1.0
setproctitle==1.1.10
simpleaudio==1.0.4
simpleline==1.7
six==1.15.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
#sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.0.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.1
srsly==2.4.2
stem==1.8.0
stemming==1.0.1
streamlit==1.2.0
#Tempita==0.5.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.13
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==20.3.0
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
uritemplate==4.1.1
urllib3==1.26.14
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.14.1
wsproto==1.2.0
xgboost==1.7.4
zipp==3.14.0
zope.event==4.2.0
zope.interface==5.1.0
six
psycopg2
